[MOD] Move PWA handling to its own hook
